pyproject.toml
[project]
name = "propostas-mcp-server"
version = "1.0.0"
description = "MCP Server for Architecture Proposal Generation System using AWS Strands Agents"
requires-python = ">=3.10"
dependencies = [
    "strands-agents>=1.29.0",
    "mcp>=1.1.2",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "notion-client>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
